+++ backend/common/src/main/java/ai/verta/modeldb/common/authservice/RoleService.java
@@ -34,7 +34,9 @@ public interface RoleService {
       List<String> entityIds, ModelDBServiceResourceTypes modelDBServiceResourceTypes);
 
   GetResourcesResponseItem getEntityResource(
-          Optional<String> entityId, Optional<String> entityName, Optional<String> workspaceName, ModelDBServiceResourceTypes modelDBServiceResourceTypes);
+          Optional<String> entityId, Optional<String> workspaceName, ModelDBServiceResourceTypes modelDBServiceResourceTypes);
+
+  List<GetResourcesResponseItem> getEntityResourcesByName(Optional<String> entityName, Optional<String> workspaceName, ModelDBServiceResourceTypes modelDBServiceResourceTypes);
 
   GeneratedMessageV3 getOrgById(String orgId);
 
@@ -94,6 +96,13 @@ public interface RoleService {
 
   void createRoleBinding(
           Role role,
+          CollaboratorBase collaborator,
+          String resourceId,
+          ModelDBServiceResourceTypes modelDBServiceResourceTypes);
+
+  void createRoleBinding(
+          String roleName,
+          RoleScope roleBindingScope,
           CollaboratorBase collaborator,
           String resourceId,
           ModelDBServiceResourceTypes modelDBServiceResourceTypes);