+++ backend/src/main/java/ai/verta/modeldb/reconcilers/ReconcilerInitializer.java
@@ -1,12 +1,13 @@
 package ai.verta.modeldb.reconcilers;
 
 import ai.verta.modeldb.ServiceSet;
+import ai.verta.modeldb.common.reconcilers.ReconcilerConfig;
 import ai.verta.modeldb.config.Config;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-public class ReconcilerUtils {
-  private static final Logger LOGGER = LogManager.getLogger(ReconcilerUtils.class);
+public class ReconcilerInitializer {
+  private static final Logger LOGGER = LogManager.getLogger(ReconcilerInitializer.class);
   public static SoftDeleteProjects softDeleteProjects;
   public static SoftDeleteExperiments softDeleteExperiments;
   public static SoftDeleteExperimentRuns softDeleteExperimentRuns;