+++ backend/src/main/java/ai/verta/modeldb/reconcilers/SoftDeleteExperimentRuns.java
@@ -3,6 +3,9 @@ package ai.verta.modeldb.reconcilers;
 import ai.verta.common.ModelDBResourceEnum;
 import ai.verta.modeldb.ModelDBConstants;
 import ai.verta.modeldb.authservice.RoleService;
+import ai.verta.modeldb.common.reconcilers.ReconcileResult;
+import ai.verta.modeldb.common.reconcilers.Reconciler;
+import ai.verta.modeldb.common.reconcilers.ReconcilerConfig;
 import ai.verta.modeldb.entities.CommentEntity;
 import ai.verta.modeldb.entities.ExperimentRunEntity;
 import ai.verta.modeldb.utils.ModelDBHibernateUtil;
@@ -42,7 +45,7 @@ public class SoftDeleteExperimentRuns extends Reconciler<String> {
   }
 
   @Override
-  protected void reconcile(Set<String> ids) {
+  protected ReconcileResult reconcile(Set<String> ids) {
     LOGGER.debug("Reconciling experiment runs " + ids.toString());
 
     deleteRoleBindings(ids);
@@ -70,9 +73,9 @@ public class SoftDeleteExperimentRuns extends Reconciler<String> {
         session.delete(experimentRunEntity);
         transaction.commit();
       }
-    } catch (Exception ex) {
-      LOGGER.error("reconcile: ", ex);
     }
+
+    return new ReconcileResult();
   }
 
   private void deleteRoleBindings(Set<String> ids) {