+++ backend/src/main/java/ai/verta/modeldb/reconcilers/SoftDeleteProjects.java
@@ -2,6 +2,9 @@ package ai.verta.modeldb.reconcilers;
 
 import ai.verta.common.ModelDBResourceEnum;
 import ai.verta.modeldb.authservice.RoleService;
+import ai.verta.modeldb.common.reconcilers.ReconcileResult;
+import ai.verta.modeldb.common.reconcilers.Reconciler;
+import ai.verta.modeldb.common.reconcilers.ReconcilerConfig;
 import ai.verta.modeldb.entities.ExperimentEntity;
 import ai.verta.modeldb.entities.ProjectEntity;
 import ai.verta.modeldb.utils.ModelDBHibernateUtil;
@@ -41,7 +44,7 @@ public class SoftDeleteProjects extends Reconciler<String> {
   }
 
   @Override
-  protected void reconcile(Set<String> ids) {
+  protected ReconcileResult reconcile(Set<String> ids) {
     LOGGER.debug("Reconciling projects " + ids.toString());
 
     roleService.deleteEntityResourcesWithServiceUser(
@@ -72,8 +75,8 @@ public class SoftDeleteProjects extends Reconciler<String> {
         session.delete(projectEntity);
         transaction.commit();
       }
-    } catch (Exception ex) {
-      LOGGER.error("reconcile: ", ex);
     }
+
+    return new ReconcileResult();
   }
 }