+++ client/verta/verta/registry/_entities/modelversions.py
@@ -4,8 +4,8 @@ from __future__ import print_function
 
 import copy
 
-from .._protos.public.registry import RegistryService_pb2 as _RegistryService
-from .._internal_utils import _utils
+from ..._protos.public.registry import RegistryService_pb2 as _RegistryService
+from ..._internal_utils import _utils
 
 from .modelversion import RegisteredModelVersion
 
@@ -20,6 +20,7 @@ class RegisteredModelVersions(_utils.LazyList):
         'time_created',
         'time_updated',
         'labels',
+        'stage',
     }
 
     def __init__(self, conn, conf):